Extracted reusable code from searchdialog.py
