+++ picard/ui/searchdialog/__init__.py
@@ -20,11 +20,7 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-
-from collections import (
-    OrderedDict,
-    namedtuple,
-)
+from collections import namedtuple
 
 from PyQt5 import (
     QtCore,
@@ -32,60 +28,22 @@ from PyQt5 import (
     QtNetwork,
     QtWidgets,
 )
-from PyQt5.QtCore import pyqtSignal
 
-from picard import (
-    config,
-    log,
-)
+from picard import config
 from picard.util import (
     icontheme,
-    natsort,
     restore_method,
-    throttle,
 )
 
-from picard.ui import PicardDialog
+from picard.ui.tablebaseddialog import TableBasedDialog
 from picard.ui.util import StandardButton
-
-
-class ResultTable(QtWidgets.QTableWidget):
-
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.setSelectionMode(
-            QtWidgets.QAbstractItemView.ExtendedSelection)
-        self.setSelectionBehavior(
-            QtWidgets.QAbstractItemView.SelectRows)
-        self.setEditTriggers(
-            QtWidgets.QAbstractItemView.NoEditTriggers)
-        self.horizontalHeader().setStretchLastSection(True)
-        self.horizontalHeader().setSectionResizeMode(
-            QtWidgets.QHeaderView.Stretch)
-        self.horizontalHeader().setSectionResizeMode(
-            QtWidgets.QHeaderView.Interactive)
-
-        @throttle(1000)  # only emit scrolled signal once per second
-        def emit_scrolled(x):
-            parent.scrolled.emit()
-        self.horizontalScrollBar().valueChanged.connect(emit_scrolled)
-        self.verticalScrollBar().valueChanged.connect(emit_scrolled)
-        self.setHorizontalScrollMode(QtWidgets.QAbstractItemView.ScrollPerPixel)
-
-    def prepare(self, headers):
-        self.clear()
-        self.setColumnCount(len(headers))
-        self.setHorizontalHeaderLabels(headers)
-        self.setRowCount(0)
-        self.setSortingEnabled(False)
 
 
 class SearchBox(QtWidgets.QWidget):
 
     def __init__(self, parent):
         super().__init__(parent)
-        self.search_action = QtWidgets.QAction(icontheme.lookup('system-search'),
-                                               _("Search"), self)
+        self.search_action = QtWidgets.QAction(icontheme.lookup('system-search'), _("Search"), self)
         self.search_action.setEnabled(False)
         self.search_action.triggered.connect(self.search)
         self.setupUi()
@@ -171,70 +129,21 @@ class SearchBox(QtWidgets.QWidget):
 Retry = namedtuple("Retry", ["function", "query"])
 
 
-class SortableTableWidgetItem(QtWidgets.QTableWidgetItem):
-
-    def __init__(self, sort_key):
-        super().__init__()
-        self.sort_key = sort_key
-
-    def __lt__(self, other):
-        return self.sort_key < other.sort_key
-
-
-class SearchDialog(PicardDialog):
-
-    defaultsize = QtCore.QSize(720, 360)
-    autorestore = False
-    scrolled = pyqtSignal()
+class SearchDialog(TableBasedDialog):
+    accept_button_title = ""
 
     def __init__(self, parent, accept_button_title, show_search=True, search_type=None):
-        super().__init__(parent)
+        self.accept_button_title = accept_button_title
         self.search_results = []
         self.show_search = show_search
         self.search_type = search_type
         self.search_box = None
-        self.setupUi(accept_button_title)
-        self.restore_state()
-        # self.columns has to be an ordered dict, with column name as keys, and
-        # matching label as values
-        self.columns = None
-        self.sorting_enabled = True
-        self.create_table()
-        self.finished.connect(self.save_state)
+        super().__init__(parent)
 
-    @property
-    def columns(self):
-        return self.__columns
+    def get_value_for_row_id(self, row, value):
+        return row
 
-    @columns.setter
-    def columns(self, list_of_tuples):
-        if not list_of_tuples:
-            list_of_tuples = []
-        self.__columns = OrderedDict(list_of_tuples)
-        self.__colkeys = list(self.columns.keys())
-
-    @property
-    def table_headers(self):
-        return list(self.columns.values())
-
-    def colpos(self, colname):
-        return self.__colkeys.index(colname)
-
-    def set_table_item(self, row, colname, obj, key, default="", sortkey=None):
-        # QVariant remembers the original type of the data
-        # matching comparison operator will be used when sorting
-        # get() will return a string, force conversion if asked to
-        value = obj.get(key, default)
-        if sortkey is None:
-            sortkey = natsort.natkey(value)
-        item = SortableTableWidgetItem(sortkey)
-        item.setData(QtCore.Qt.DisplayRole, value)
-        pos = self.colpos(colname)
-        if pos == 0:
-            item.setData(QtCore.Qt.UserRole, row)
-        self.table.setItem(row, pos, item)
-
-    def setupUi(self, accept_button_title):
+    def setupUi(self):
         self.verticalLayout = QtWidgets.QVBoxLayout(self)
         self.verticalLayout.setObjectName("vertical_layout")
         if self.show_search:
@@ -257,7 +166,7 @@ class SearchDialog(PicardDialog):
                 QtWidgets.QDialogButtonBox.ActionRole)
             self.search_browser_button.clicked.connect(self.search_browser)
         self.accept_button = QtWidgets.QPushButton(
-            accept_button_title,
+            self.accept_button_title,
             self.buttonBox)
         self.accept_button.setEnabled(False)
         self.buttonBox.addButton(
@@ -269,19 +178,6 @@ class SearchDialog(PicardDialog):
         self.buttonBox.accepted.connect(self.accept)
         self.buttonBox.rejected.connect(self.reject)
         self.verticalLayout.addWidget(self.buttonBox)
-
-    def add_widget_to_center_layout(self, widget):
-        """Update center widget with new child. If child widget exists,
-        schedule it for deletion."""
-        widget_item = self.center_layout.takeAt(0)
-        if widget_item:
-            current_widget = widget_item.widget()
-            current_widget.hide()
-            self.center_layout.removeWidget(current_widget)
-            if current_widget != self.table:
-                current_widget.deleteLater()
-        self.center_layout.addWidget(widget)
-        widget.show()
 
     def show_progress(self):
         progress_widget = QtWidgets.QWidget(self)
@@ -328,22 +224,6 @@ class SearchDialog(PicardDialog):
         error_widget.setLayout(layout)
         self.add_widget_to_center_layout(error_widget)
 
-    def create_table(self):
-        self.table = ResultTable(self)
-        self.table.verticalHeader().setDefaultSectionSize(100)
-        self.table.setSortingEnabled(False)
-        self.table.cellDoubleClicked.connect(self.accept)
-        self.table.hide()
-
-        def enable_accept_button():
-            self.accept_button.setEnabled(True)
-        self.table.itemSelectionChanged.connect(
-            enable_accept_button)
-
-    def prepare_table(self):
-        self.table.prepare(self.table_headers)
-        self.restore_table_header_state()
-
     def show_table(self, sort_column=None, sort_order=QtCore.Qt.DescendingOrder):
         self.add_widget_to_center_layout(self.table)
         self.table.horizontalHeader().setSortIndicatorShown(self.sorting_enabled)
@@ -385,27 +265,9 @@ class SearchDialog(PicardDialog):
 
     @restore_method
     def restore_state(self):
-        self.restore_geometry()
+        super().restore_state()
         if self.show_search:
             self.search_box.restore_checkbox_state()
-
-    @restore_method
-    def restore_table_header_state(self):
-        header = self.table.horizontalHeader()
-        state = config.persist[self.dialog_header_state]
-        if state:
-            header.restoreState(state)
-        header.setSectionResizeMode(QtWidgets.QHeaderView.Interactive)
-        log.debug("restore_state: %s" % self.dialog_header_state)
-
-    def save_state(self):
-        if self.table:
-            self.save_table_header_state()
-
-    def save_table_header_state(self):
-        state = self.table.horizontalHeader().saveState()
-        config.persist[self.dialog_header_state] = state
-        log.debug("save_state: %s" % self.dialog_header_state)
 
     def search_box_text(self, text):
         if self.search_box:
